move common code into test fixtures
