+++ system-tests/azure-data-factory-tests/src/test/java/org/eclipse/dataspaceconnector/system/tests/local/AzureDataFactoryTransferIntegrationTest.java
@@ -14,18 +14,11 @@
 
 package org.eclipse.dataspaceconnector.system.tests.local;
 
-import io.restassured.specification.RequestSpecification;
-import org.eclipse.dataspaceconnector.azure.blob.core.AzureBlobStoreSchema;
 import org.eclipse.dataspaceconnector.azure.blob.core.api.BlobStoreApiImpl;
 import org.eclipse.dataspaceconnector.azure.testfixtures.annotations.AzureDataFactoryIntegrationTest;
 import org.eclipse.dataspaceconnector.common.testfixtures.TestUtils;
 import org.eclipse.dataspaceconnector.core.security.azure.AzureVault;
 import org.eclipse.dataspaceconnector.junit.launcher.EdcRuntimeExtension;
-import org.eclipse.dataspaceconnector.policy.model.Action;
-import org.eclipse.dataspaceconnector.policy.model.Permission;
-import org.eclipse.dataspaceconnector.policy.model.Policy;
-import org.eclipse.dataspaceconnector.policy.model.PolicyType;
-import org.eclipse.dataspaceconnector.spi.asset.AssetSelectorExpression;
 import org.eclipse.dataspaceconnector.spi.monitor.ConsoleMonitor;
 import org.eclipse.dataspaceconnector.system.tests.utils.TransferSimulationUtils;
 import org.jetbrains.annotations.NotNull;
@@ -45,13 +38,13 @@ import java.util.Map;
 import java.util.Properties;
 import java.util.UUID;
 
-import static io.restassured.RestAssured.given;
-import static io.restassured.http.ContentType.JSON;
-import static java.lang.String.format;
 import static java.lang.String.valueOf;
 import static java.lang.System.getenv;
 import static java.nio.charset.StandardCharsets.UTF_8;
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.eclipse.dataspaceconnector.system.tests.local.BlobTransferUtils.createAsset;
+import static org.eclipse.dataspaceconnector.system.tests.local.BlobTransferUtils.createContractDefinition;
+import static org.eclipse.dataspaceconnector.system.tests.local.BlobTransferUtils.createPolicy;
 import static org.eclipse.dataspaceconnector.system.tests.local.TransferLocalSimulation.CONSUMER_CONNECTOR_PATH;
 import static org.eclipse.dataspaceconnector.system.tests.local.TransferLocalSimulation.CONSUMER_CONNECTOR_PORT;
 import static org.eclipse.dataspaceconnector.system.tests.local.TransferLocalSimulation.CONSUMER_IDS_API;
@@ -59,7 +52,6 @@ import static org.eclipse.dataspaceconnector.system.tests.local.TransferLocalSim
 import static org.eclipse.dataspaceconnector.system.tests.local.TransferLocalSimulation.CONSUMER_MANAGEMENT_PATH;
 import static org.eclipse.dataspaceconnector.system.tests.local.TransferLocalSimulation.CONSUMER_MANAGEMENT_PORT;
 import static org.eclipse.dataspaceconnector.system.tests.local.TransferLocalSimulation.IDS_PATH;
-import static org.eclipse.dataspaceconnector.system.tests.local.TransferLocalSimulation.PROVIDER_CONNECTOR_MANAGEMENT_URL;
 import static org.eclipse.dataspaceconnector.system.tests.local.TransferLocalSimulation.PROVIDER_CONNECTOR_PATH;
 import static org.eclipse.dataspaceconnector.system.tests.local.TransferLocalSimulation.PROVIDER_CONNECTOR_PORT;
 import static org.eclipse.dataspaceconnector.system.tests.local.TransferLocalSimulation.PROVIDER_IDS_API;
@@ -67,20 +59,17 @@ import static org.eclipse.dataspaceconnector.system.tests.local.TransferLocalSim
 import static org.eclipse.dataspaceconnector.system.tests.local.TransferLocalSimulation.PROVIDER_MANAGEMENT_PATH;
 import static org.eclipse.dataspaceconnector.system.tests.local.TransferLocalSimulation.PROVIDER_MANAGEMENT_PORT;
 import static org.eclipse.dataspaceconnector.system.tests.utils.GatlingUtils.runGatling;
-import static org.eclipse.dataspaceconnector.system.tests.utils.TransferSimulationUtils.PROVIDER_ASSET_ID;
+import static org.eclipse.dataspaceconnector.system.tests.utils.TransferSimulationUtils.PROVIDER_ASSET_FILE;
 
 @AzureDataFactoryIntegrationTest
 public class AzureDataFactoryTransferIntegrationTest {
 
     private static final List<Runnable> containerCleanup = new ArrayList<>();
-    private static final String ASSETS_PATH = "/assets";
-    private static final String POLICIES_PATH = "/policies";
     private static final String EDC_FS_CONFIG = "edc.fs.config";
     private static final String EDC_VAULT_NAME = "edc.vault.name";
     private static final String EDC_VAULT_CLIENT_ID = "edc.vault.clientid";
     private static final String EDC_VAULT_TENANT_ID = "edc.vault.tenantid";
     private static final String EDC_VAULT_CLIENT_SECRET = "edc.vault.clientsecret";
-    private static final String CONTRACT_DEFINITIONS_PATH = "/contractdefinitions";
     private static final String PROVIDER_CONTAINER_NAME = UUID.randomUUID().toString();
     private static final String KEY_VAULT_NAME = runtimeSettingsProperties().getProperty("test.key.vault.name");
     private static final String AZURE_TENANT_ID = getenv("AZURE_TENANT_ID");
@@ -89,7 +78,6 @@ public class AzureDataFactoryTransferIntegrationTest {
     private static final String PROVIDER_STORAGE_ACCOUNT_NAME = runtimeSettingsProperties().getProperty("test.provider.storage.name");
     private static final String CONSUMER_STORAGE_ACCOUNT_NAME = runtimeSettingsProperties().getProperty("test.consumer.storage.name");
     private static final String BLOB_STORE_ENDPOINT_TEMPLATE = "https://%s.blob.core.windows.net";
-    private static final String KEY_VAULT_ENDPOINT_TEMPLATE = "https://%s.vault.azure.net";
 
     @RegisterExtension
     private static final EdcRuntimeExtension consumer = new EdcRuntimeExtension(
@@ -153,12 +141,12 @@ public class AzureDataFactoryTransferIntegrationTest {
         var blobContent = "AzureDataFactoryTransferIntegrationTest-" + UUID.randomUUID();
 
         blobStoreApi.createContainer(PROVIDER_STORAGE_ACCOUNT_NAME, PROVIDER_CONTAINER_NAME);
-        blobStoreApi.putBlob(PROVIDER_STORAGE_ACCOUNT_NAME, PROVIDER_CONTAINER_NAME, PROVIDER_ASSET_ID, blobContent.getBytes(UTF_8));
+        blobStoreApi.putBlob(PROVIDER_STORAGE_ACCOUNT_NAME, PROVIDER_CONTAINER_NAME, PROVIDER_ASSET_FILE, blobContent.getBytes(UTF_8));
         // Add for cleanup
         containerCleanup.add(() -> blobStoreApi.deleteContainer(PROVIDER_STORAGE_ACCOUNT_NAME, PROVIDER_CONTAINER_NAME));
 
         // Seed data to provider
-        createAsset();
+        createAsset(PROVIDER_STORAGE_ACCOUNT_NAME, PROVIDER_CONTAINER_NAME);
         var policyId = createPolicy();
         createContractDefinition(policyId);
 
@@ -172,9 +160,9 @@ public class AzureDataFactoryTransferIntegrationTest {
         // Add for cleanup
         containerCleanup.add(() -> blobStoreApi.deleteContainer(CONSUMER_STORAGE_ACCOUNT_NAME, provisionedContainerName));
 
-        var actualBlobContent = blobStoreApi.getBlob(CONSUMER_STORAGE_ACCOUNT_NAME, provisionedContainerName, PROVIDER_ASSET_ID);
+        var actualBlobContent = blobStoreApi.getBlob(CONSUMER_STORAGE_ACCOUNT_NAME, provisionedContainerName, PROVIDER_ASSET_FILE);
         assertThat(actualBlobContent.length)
-                .withFailMessage("Destination blob %s not created", PROVIDER_ASSET_ID)
+                .withFailMessage("Destination blob %s not created", PROVIDER_ASSET_FILE)
                 .isGreaterThan(0);
         assertThat(new String(actualBlobContent))
                 .withFailMessage("Transferred file contents are not same as the source file")
@@ -197,77 +185,5 @@ public class AzureDataFactoryTransferIntegrationTest {
         } catch (IOException e) {
             throw new RuntimeException("Error in loading runtime settings properties", e);
         }
-    }
-
-    private void createAsset() {
-        var asset = Map.of(
-                "asset", Map.of(
-                        "properties", Map.of(
-                                "asset:prop:name", PROVIDER_ASSET_ID,
-                                "asset:prop:contenttype", "text/plain",
-                                "asset:prop:version", "1.0",
-                                "asset:prop:id", PROVIDER_ASSET_ID,
-                                "type", "AzureStorage"
-                        )
-                ),
-                "dataAddress", Map.of(
-                        "properties", Map.of(
-                                "type", AzureBlobStoreSchema.TYPE,
-                                AzureBlobStoreSchema.ACCOUNT_NAME, PROVIDER_STORAGE_ACCOUNT_NAME,
-                                AzureBlobStoreSchema.CONTAINER_NAME, PROVIDER_CONTAINER_NAME,
-                                AzureBlobStoreSchema.BLOB_NAME, PROVIDER_ASSET_ID,
-                                "keyName", format("%s-key1", PROVIDER_STORAGE_ACCOUNT_NAME)
-                        )
-                )
-        );
-
-        seedProviderData(ASSETS_PATH, asset);
-    }
-
-    private String createPolicy() {
-        var policy = Policy.Builder.newInstance()
-                .permission(Permission.Builder.newInstance()
-                        .target(PROVIDER_ASSET_ID)
-                        .action(Action.Builder.newInstance().type("USE").build())
-                        .build())
-                .type(PolicyType.SET)
-                .build();
-
-        seedProviderData(POLICIES_PATH, policy);
-
-        return policy.getUid();
-    }
-
-    private void createContractDefinition(String policyId) {
-
-        var criteria = AssetSelectorExpression.Builder.newInstance()
-                .constraint("asset:prop:id",
-                        "=",
-                        PROVIDER_ASSET_ID)
-                .build();
-
-        var contractDefinition = Map.of(
-                "id", "1",
-                "accessPolicyId", policyId,
-                "contractPolicyId", policyId,
-                "criteria", criteria.getCriteria()
-        );
-
-        seedProviderData(CONTRACT_DEFINITIONS_PATH, contractDefinition);
-    }
-
-    private void seedProviderData(String path, Object requestBody) {
-        givenProviderBaseRequest()
-                .contentType(JSON)
-                .body(requestBody)
-                .when()
-                .post(path)
-                .then()
-                .statusCode(204);
-    }
-
-    private RequestSpecification givenProviderBaseRequest() {
-        return given()
-                .baseUri(PROVIDER_CONNECTOR_MANAGEMENT_URL + PROVIDER_MANAGEMENT_PATH);
     }
 }
